refactor(angular): migrate tabs to common
